+++ src/app/settings/settings.component.ts
@@ -30,7 +30,7 @@ const blankSettings: SettingsType = {
   mission: '', event: '', eventNotes: '', opPeriod: '',
   opPeriodStart: new Date(0), opPeriodEnd: new Date(0),
   application: '', version: '', debugMode: false,
-  defLat: 0, defLng: 0, defPlusCode: '', w3wLocale: '', allowManualPinDrops: false,
+  defLat: 0, defLng: 0, allowManualPinDrops: false,
   googleGeocodingApiKey: '',
   showDD: true, showDDM: true, showDMS: true, showMGRS: true, showUTM: true, showMaidenhead: true,
   maplibre: { defZoom: 15, markerScheme: '', overviewDifference: 5, overviewMinZoom: 5, overviewMaxZoom: 16 },