+++ src/app/shared/index.ts
@@ -2,7 +2,7 @@
 export { Utility } from "./utility"
 export { AlertsComponent } from "./alerts/alerts.component"
 export { BackToTopComponent } from "./back-to-top/back-to-top.component"
-export { DisclosureComponent } from "./disclosure/disclosure.component"
+export { SectionComponent } from "./section/section.component"
 export { HeaderComponent } from "./header/header.component"
 export { PageComponent } from "./page/page.component"
 export { IconsComponent } from "./icons/icons.component"